changed from decimal to float

+++ src/leeger/model/Matchup.py
@@ -1,5 +1,4 @@
 from dataclasses import dataclass
-from decimal import Decimal
 
 from src.leeger.model.abstract.UniqueId import UniqueId
 
@@ -8,5 +7,5 @@ from src.leeger.model.abstract.UniqueId import UniqueId
 class Matchup(UniqueId):
     teamAId: str
     teamBId: str
-    teamAScore: Decimal
-    teamBScore: Decimal
+    teamAScore: float
+    teamBScore: float
